+++ src/pymonzo/monzo_api.py
@@ -15,7 +15,7 @@ from six.moves.urllib.parse import urljoin
 
 from pymonzo.api_objects import MonzoAccount, MonzoBalance, MonzoTransaction, MonzoToken
 from pymonzo import config
-from pymonzo.exceptions import MonzoAPIException
+from pymonzo.exceptions import MonzoAPIException, UnableToRefreshTokenException
 from pymonzo.utils import CommonMixin
 from pymonzo.logger import logging
 
@@ -170,8 +170,7 @@ class MonzoAPI(CommonMixin):
         Official docs:
             https://monzo.com/docs/#refreshing-access
 
-        :returns: OAuth 2 access token
-        :rtype: dict
+        :raises UnableToRefreshTokenException: when token couldn't be refreshed
         """
         self._logger.debug('refreshing oauth token')
         url = urljoin(self.api_url, '/oauth2/token')
@@ -190,11 +189,15 @@ class MonzoAPI(CommonMixin):
         except ValueError as e:
             self._logger.debug('Error getting token -> {}'.format(e.message))
             self._logger.debug('Token Data: {}'.format(token))
-            return None
+            raise UnableToRefreshTokenException("Unable to refresh the token: {}".format(token))
 
+        if 'error' in token:
+            raise UnableToRefreshTokenException(
+                "Unable to refresh the token: {}".format(token)
+            )
+
+        self._token = token
         self._save_token_on_disk(token)
-
-        return token
 
     def _get_response(self, method, endpoint, params=None, data=None):
         """
@@ -221,10 +224,11 @@ class MonzoAPI(CommonMixin):
             self._logger.debug('response status code: {}'.format(response.status_code))
             if response.status_code == 401:
                 raise TokenExpiredError()
+
         except TokenExpiredError:
             # For some reason 'requests-oauthlib' automatic token refreshing
             # doesn't work so we do it here semi-manually
-            self._token = self._refresh_oath_token()
+            self._refresh_oath_token()
 
             self._session = OAuth2Session(
                 client_id=self._client_id,
@@ -238,7 +242,7 @@ class MonzoAPI(CommonMixin):
 
         if response.status_code != requests.codes.ok:
             raise MonzoAPIException(
-                "Something wrong happened: {}".format(response.json())
+                "Something went wrong: {}".format(response.json())
             )
 
         return response